+++ NEWS.md
@@ -1,5 +1,7 @@
 # Changes in version 1.1-1.1 (xx/xx/2016)
 
+* impoved grapical parameters for the grouped matrix plot.
+
 # Changes in version 1.1-1 (04/10/2016)
 
 * plot method ordered now passes control arguments on correctly.